requirements.txt

discord
jishaku
motor
python-dotenv
